Update links to ROS 2 documentation.
I am not sure if these index links used to work previously but they
aren't valid currently.

I had started down the road of updating the tutorial to current ROS 2
distributions when I realized that this tutorial might be restricted to
Gazebo Classic and I'm not sure how much of an overhaul it might need
for Humble and Iron.

Happy to continue updates in follow-up PRs if it makes sense to do so.